requirements.txt
openai
Pillow
requests
pypdfium2
pymupdf
numpy
scipy
tiktoken
pytesseract
opencv-python
matplotlib

